fix: rename react folder to react-components to prevent Vite error
Renaming the `react` folder to `react-components` avoids an issue where Vite
may mistakenly interpret the folder as a module, causing erratic errors.
