+++ assets/copy/react-components/index.js
