+++ assets/copy/react-components/simple.jsx
