+++ live_react_examples/.gitignore
@@ -27,7 +27,7 @@ live_react_examples-*.tar
 
 # Ignore assets that are produced by build tools.
 /priv/static/assets/
-/priv/react/
+/priv/react-components/
 
 # Ignore digested assets cache.
 /priv/static/cache_manifest.json
