+++ live_react_examples/assets/js/server.js
@@ -1,4 +1,4 @@
-import components from "../react";
+import components from "../react-components";
 import { getRender } from "live_react/server";
 
 export const render = getRender(components);
