+++ live_react_examples/assets/package.json
@@ -4,7 +4,7 @@
   "scripts": {
     "dev": "vite --host",
     "build": "tsc && vite build",
-    "build-server": "tsc && vite build --ssr js/server.js --out-dir ../priv/react --minify esbuild && echo '{\"type\": \"module\" } ' > ../priv/react/package.json"
+    "build-server": "tsc && vite build --ssr js/server.js --out-dir ../priv/react-components --minify esbuild && echo '{\"type\": \"module\" } ' > ../priv/react-components/package.json"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.7",
