+++ live_react_examples/assets/react-components/components/lazy-component.jsx
