+++ live_react_examples/assets/react-components/counter.jsx
