+++ live_react_examples/assets/react-components/github-code.jsx
