+++ live_react_examples/assets/react-components/index.jsx
