+++ live_react_examples/assets/react-components/lazy.jsx
