+++ live_react_examples/assets/react-components/log-list.jsx
