+++ live_react_examples/assets/react-components/simple-props.jsx
