+++ live_react_examples/assets/react-components/ssr.jsx
