+++ live_react_examples/assets/react-components/typescript.tsx
